Added register request model
