+++ src/main/java/com/auctionapp/api/model/dto/RegisterRequest.java
@@ -1,6 +1,4 @@
-package com.auctionapp.api.model;
-
-import java.sql.Timestamp;
+package com.auctionapp.api.model.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
@@ -9,13 +7,9 @@ import lombok.NoArgsConstructor;
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
-public class UserDto {
-  private Long uuid;
+public class RegisterRequest {
   private String firstName;
   private String lastName;
-  private String username;
   private String email;
   private String password;
-  private Timestamp createdAt;
-  private Timestamp updatedAt;
 }
